fix(mods): avoid eager React resolution for non-React mods
Letta Code (agent-c2adbf5c-8419-4211-8cd8-3740db164974)

Only prepare runtime dependency symlinks after a mod has been transpiled and scanned for actual runtime React imports, so AppImage-style runtimes without unpacked React no longer fail plain TypeScript mods during reload.
