+++ scripts/source-file-size-baseline.json
@@ -26,8 +26,8 @@
   "src/hooks/integration.test.ts": 1147,
   "src/index.ts": 2803,
   "src/mods/learning-harness.ts": 2434,
-  "src/mods/mod-engine.test.ts": 2178,
-  "src/mods/mod-engine.ts": 1879,
+  "src/mods/mod-engine.test.ts": 2132,
+  "src/mods/mod-engine.ts": 1838,
   "src/mods/package-installer.test.ts": 1248,
   "src/mods/package-installer.ts": 1201,
   "src/mods/package-registry.ts": 1033,
