+++ src/mods/mod-engine.ts
@@ -1,13 +1,10 @@
 import { createHash } from "node:crypto";
 import {
   existsSync,
-  lstatSync,
   mkdirSync,
   readdirSync,
   readFileSync,
-  readlinkSync,
   statSync,
-  symlinkSync,
   unlinkSync,
   writeFileSync,
 } from "node:fs";
@@ -57,6 +54,7 @@ import {
   unregisterModPermission,
   unregisterModPermissionsForOwner,
 } from "@/mods/permission-registry";
+import { ensureRuntimeDependenciesForModCache } from "@/mods/runtime-dependencies";
 import {
   getModToolDefinition,
   type ModToolDefinition,
@@ -109,6 +107,7 @@ export const LEGACY_GLOBAL_EXTENSIONS_DIRECTORY =
 export const MOD_CACHE_DIRECTORY = getModCacheDirectory();
 
 const requireFromRuntime = createRequire(import.meta.url);
+let resolveRuntimePackageDirectory = getRuntimePackageDirectory;
 
 export type LettaModDisposer = () => void;
 
@@ -413,54 +412,10 @@ function getRuntimePackageDirectory(packageName: string): string {
   );
 }
 
-function normalizeRuntimeDependencyPath(value: string): string {
-  const normalized = path.normalize(value).replace(/^\\\\\?\\/, "");
-  return process.platform === "win32" ? normalized.toLowerCase() : normalized;
-}
-
-function ensureRuntimeDependencySymlink(
-  cacheDirectory: string,
-  packageName: string,
+export function __testOverrideRuntimePackageDirectoryResolver(
+  resolver: ((packageName: string) => string) | null,
 ): void {
-  const nodeModulesDirectory = path.join(cacheDirectory, "node_modules");
-  const linkPath = path.join(nodeModulesDirectory, packageName);
-  const packageDirectory = path.resolve(
-    getRuntimePackageDirectory(packageName),
-  );
-
-  mkdirSync(nodeModulesDirectory, { recursive: true });
-  try {
-    const stats = lstatSync(linkPath);
-    if (!stats.isSymbolicLink()) return;
-
-    const existingTarget = readlinkSync(linkPath);
-    const resolvedTarget = path.resolve(nodeModulesDirectory, existingTarget);
-    if (
-      normalizeRuntimeDependencyPath(resolvedTarget) ===
-      normalizeRuntimeDependencyPath(packageDirectory)
-    ) {
-      return;
-    }
-
-    unlinkSync(linkPath);
-  } catch (error) {
-    if ((error as NodeJS.ErrnoException).code !== "ENOENT") throw error;
-  }
-
-  try {
-    symlinkSync(
-      packageDirectory,
-      linkPath,
-      process.platform === "win32" ? "junction" : "dir",
-    );
-  } catch (error) {
-    if ((error as NodeJS.ErrnoException).code !== "EEXIST") throw error;
-  }
-}
-
-function ensureModCache(cacheDirectory: string): void {
-  mkdirSync(cacheDirectory, { recursive: true });
-  ensureRuntimeDependencySymlink(cacheDirectory, "react");
+  resolveRuntimePackageDirectory = resolver ?? getRuntimePackageDirectory;
 }
 
 function isPathInsideOrEqual(childPath: string, parentPath: string): boolean {
@@ -533,12 +488,6 @@ function createImportableModPath(
 ): string {
   const importCacheDirectory =
     getManagedPackageImportCacheDirectory(modPath, source) ?? cacheDirectory;
-  if (importCacheDirectory === cacheDirectory) {
-    ensureModCache(importCacheDirectory);
-  } else {
-    mkdirSync(importCacheDirectory, { recursive: true });
-  }
-
   const sourceText = readFileSync(modPath, "utf8");
   const hash = createHash("sha256")
     .update(sourceText)
@@ -546,6 +495,16 @@ function createImportableModPath(
     .slice(0, 16);
   const fileExtension = path.extname(modPath);
   const importableSource = prepareModForImport(modPath, sourceText);
+
+  if (importCacheDirectory === cacheDirectory) {
+    ensureRuntimeDependenciesForModCache(
+      importCacheDirectory,
+      importableSource,
+      resolveRuntimePackageDirectory,
+    );
+  } else {
+    mkdirSync(importCacheDirectory, { recursive: true });
+  }
   const baseName = path
     .basename(modPath, fileExtension)
     .replace(/[^a-zA-Z0-9_-]/g, "-");
